third_party/WebKit/LayoutTests/fast/frames/xss-auditor-handles-file-urls-expected.txt

CONSOLE ERROR: line 3: The XSS Auditor blocked access to 'html;base64,PHNjcmlwdD5hbGVydCgxKTwvc2NyaXB0Pg==%27%3E' because the source code of a script was found within the request. The auditor was enabled as the server did not send an 'X-XSS-Protection' header.
